Issue #3208 - Combined migrations for miq_report_results and binary_blobs into one migration with a shared module containing the serialization methods.
