+++ vmdb/db/migrate/20150625220141_fix_serialized_reports_for_rails_four.rb
@@ -1,39 +1,51 @@
-class BinaryBlobFixSerializedReport < ActiveRecord::Migration
-  class BinaryBlob < ActiveRecord::Base
-    self.inheritance_column = :_type_disabled
-    has_many :binary_blob_parts, -> { order(:id) }
-
+class FixSerializedReportsForRailsFour < ActiveRecord::Migration
+  module Serializer
     YAML_ATTRS = [:table, :sub_table, :filter_summary, :extras, :ids, :scoped_association, :html_title, :file_name, :extras, :record_id,
                   :tl_times, :user_categories, :trend_data, :performance, :include_for_find, :report_run_time, :chart, :reserved]
 
-    def serialize_report_to_hash
-      val = binary
-
+    def serialize_report_to_hash(val)
       if val.include?("!ruby/object:MiqReport")
         val.sub!(/MiqReport/, 'Hash')
       else
-        raise "unexpected format of binary data encountered, '#{val.inspect}'"
+        raise "unexpected format of report attribute encountered, '#{val.inspect}'"
       end
-
       raw_hash = YAML.load(val)
+      raw_hash = raw_hash.last if raw_hash.kind_of?(Array) # MiqReport was serialized as an Array with 1 element in miq_report_results
       new_hash = YAML_ATTRS.each_with_object(raw_hash['attributes'].to_hash) { |k, h| h[k.to_s] = raw_hash[k.to_s] }
 
       YAML.dump(new_hash)
     end
 
-    def serialize_hash_to_report
-      val = binary
+    def serialize_hash_to_report(val, from)
       if val.starts_with?("---")
         new_hash = {'attributes' => {}}
         YAML.load(val).each do |k,v|
           YAML_ATTRS.include?(k.to_sym) ? new_hash[k.to_s] = v : new_hash['attributes'][k.to_s] = v
         end
 
-        YAML.dump(new_hash).sub(/---/, "--- !ruby/object:MiqReport")
+        if from == :miq_report_result
+          # MiqReport was serialized as an Array with 1 element in miq_report_results
+          YAML.dump([new_hash]).sub(/---\n- attributes:/, "---\n- !ruby/object:MiqReport\n  attributes:")
+        else
+          YAML.dump(new_hash).sub(/---/, "--- !ruby/object:MiqReport")
+        end
       else
         raise "unexpected format of report attribute encountered, '#{val.inspect}'"
       end
     end
+  end
+
+  class MiqReportResult < ActiveRecord::Base
+    self.inheritance_column = :_type_disabled
+
+    include Serializer
+  end
+
+  class BinaryBlob < ActiveRecord::Base
+    self.inheritance_column = :_type_disabled
+    has_many :binary_blob_parts, -> { order(:id) }
+
+    include Serializer
 
     def delete_binary
       self.md5 = self.size = self.part_size = nil
@@ -72,17 +84,29 @@ class BinaryBlobFixSerializedReport < ActiveRecord::Migration
   end
 
   def up
+    say_with_time("Converting MiqReportResult#report to a serialized hash") do
+      MiqReportResult.where('report IS NOT NULL').each do |rr|
+        rr.update_attribute(:report, rr.serialize_report_to_hash(rr.read_attribute(:report)))
+      end
+    end
+
     say_with_time("Converting BinaryBlob report results to a serialized hash") do
       BinaryBlob.where(:resource_type=>'MiqReportResult').each do |bb|
-        bb.binary = bb.serialize_report_to_hash
+        bb.binary = bb.serialize_report_to_hash(bb.binary)
       end
     end
   end
 
   def down
+    say_with_time("Converting MiqReportResult#report back to a serialized MiqReport") do
+      MiqReportResult.where('report IS NOT NULL').each do |rr|
+        rr.update_attribute(:report, rr.serialize_hash_to_report(rr.read_attribute(:report), :miq_report_result))
+      end
+    end
+
     say_with_time("Converting BinaryBlob report results back to a serialized MiqReport") do
       BinaryBlob.where(:resource_type=>'MiqReportResult').each do |bb|
-        bb.binary = bb.serialize_hash_to_report
+        bb.binary = bb.serialize_hash_to_report(bb.binary, :binary_blob)
       end
     end
   end
